* [LocTT] Update Chart and Report 20220603

+++ KhoanNhaTrang/Model/Data.cs
@@ -7,10 +7,10 @@ namespace KhoanNhaTrang.Model
     class Data
     {
         public long Id { get; set; }
-        public float flow_rate { get; set; }
-        public float fluid { get; set; }
-        public float pressure { get; set; }
-        public float wc { get; set; }
+        public double flow_rate { get; set; }
+        public double fluid { get; set; }
+        public double pressure { get; set; }
+        public double wc { get; set; }
 
         public float insert_date { get; set; }
